requirements.txt
numpy  # C-like arrays and vectorized functions
pyzmq  # 0MQ
paramiko  # SSH API in python
pyyaml  # yaml
tables  # pytables HDF5 library in Python
matplotlib  # Plotting colors
uncertainties  # Error propagation
pytest  # Tests
pyqtgraph==0.11  # Fast plotting
pyqt5==5.12  # Qt library in Python
